refactor SearchTestHelper into AbstractSearchTest
renames SearchTestHelper and moves auxiliary functions from the subclass
to the parent
